Rename config module to env
